Move search request to Content
